+++ source/computer/puzzle/PuzzleSystemGUI.cs
@@ -8,7 +8,7 @@ public class PuzzleSystemGUI : BaseSystemGUI
 {
 	public void UpdateGUIState(bool active)
 	{
-		SCG.IEnumerator<SCG.KeyValuePair<byte, Button>> it = buttonMap.GetEnumerator();
+		SCG.IEnumerator<SCG.KeyValuePair<int, Button>> it = buttonMap.GetEnumerator();
 			
 		while(it.MoveNext())
 			it.Current.Value.Disabled = !active;
@@ -19,23 +19,22 @@ public class PuzzleSystemGUI : BaseSystemGUI
 		inactivePanel.Visible = !active;
 	}
 
-  private void InitializeMaps()
-  {
-		buttonMap = new Dictionary<byte, Button>();
-		labelMap = new Dictionary<byte, Label>();
-		textureRectMap =  new Dictionary<byte, TextureRect>();
-    touchAreaMap = new Dictionary<ulong, byte>();
+	private void InitializeMaps()
+	{
+		buttonMap = new Dictionary<int, Button>();
+		labelMap = new Dictionary<int, Label>();
+		textureRectMap =  new Dictionary<int, TextureRect>();
+		touchAreaMap = new Dictionary<ulong, byte>();
 		inactivePanel = GetNode<PanelContainer>(inactivePanelNP);
-  }
+	}
 
 	private void InitializeAnswerPanel()
 	{
 		answerPanel = GetNode<PanelContainer>(answerPanelNP);
 		Node pwdc = answerPanel.GetChild(0);
-    Node child;
+		Node child;
 		
-		for(byte i = SystemGUIID.LBL_PWD_C0;
-        i <= SystemGUIID.LBL_PWD_C11; i++)
+		for(byte i = SystemGUIID.LBL_PWD_C0; i <= SystemGUIID.LBL_PWD_C11; i++)
 		{
 			child = pwdc.GetChild(i - SystemGUIID.LBL_PWD_C0);
 			labelMap.Add(i, child.GetChild<Label>(0));
@@ -45,7 +44,7 @@ public class PuzzleSystemGUI : BaseSystemGUI
 
 	private void InitializePuzzleContentPanel()
 	{
-    puzzleContentPanel = GetNode<PanelContainer>(puzzleContentPanelNP);
+		puzzleContentPanel = GetNode<PanelContainer>(puzzleContentPanelNP);
 		Node cc = puzzleContentPanel.GetChild(0);
 		Node pp = cc.GetChild(0);
 		labelMap.Add(SystemGUIID.LBL_PAGE_CONTENT, pp.GetChild<Label>(0));
@@ -60,8 +59,7 @@ public class PuzzleSystemGUI : BaseSystemGUI
 		Node tck = touchKeyboardPanel.GetChild(0);
 		Node child;
 		
-		for(byte i = SystemGUIID.BTN_KB_KEY0;
-        i <= SystemGUIID.BTN_KB_KEY_ALT; i++)
+		for(byte i = SystemGUIID.BTN_KB_KEY0; i <= SystemGUIID.BTN_KB_KEY_ALT; i++)
 		{
 			child = tck.GetChild(i - SystemGUIID.BTN_KB_KEY0);
 			textureRectMap.Add(i, child.GetChild<TextureRect>(0));
@@ -69,33 +67,30 @@ public class PuzzleSystemGUI : BaseSystemGUI
 		}
 	}
 
-  private void InitializeTouchArea()
-  {
-    Node ta = GetNode(touchAreaNP);
-    Node pa = ta.GetChild(0);
-    Node kba = ta.GetChild(1);
-    touchAreaMap.Add(pa.GetChild(0).GetInstanceId(),
-        SystemGUIID.BTN_PREV_PAGE);
-    touchAreaMap.Add(pa.GetChild(1).GetInstanceId(),
-        SystemGUIID.BTN_NEXT_PAGE);
-    Node child;
+	private void InitializeTouchArea()
+	{
+		Node ta = GetNode(touchAreaNP);
+		Node pa = ta.GetChild(0);
+		Node kba = ta.GetChild(1);
+		touchAreaMap.Add(pa.GetChild(0).GetInstanceId(), SystemGUIID.BTN_PREV_PAGE);
+		touchAreaMap.Add(pa.GetChild(1).GetInstanceId(), SystemGUIID.BTN_NEXT_PAGE);
+		Node child;
 
-    for(byte i = SystemGUIID.BTN_KB_KEY0;
-        i <= SystemGUIID.BTN_KB_KEY_ALT; i++)
-    {
-      child = kba.GetChild(i - SystemGUIID.BTN_KB_KEY0);
-      touchAreaMap.Add(child.GetInstanceId(), i);
-    }
-  }
+		for(byte i = SystemGUIID.BTN_KB_KEY0; i <= SystemGUIID.BTN_KB_KEY_ALT; i++)
+		{
+		  child = kba.GetChild(i - SystemGUIID.BTN_KB_KEY0);
+		  touchAreaMap.Add(child.GetInstanceId(), i);
+		}
+	}
 
-  public override void _EnterTree()
-  {
-    InitializeMaps();
-    InitializeAnswerPanel();
-    InitializePuzzleContentPanel();
-    InitializeTouchKeyboardPanel();
-    InitializeTouchArea();
-  }
+	public override void _EnterTree()
+	{
+		InitializeMaps();
+		InitializeAnswerPanel();
+		InitializePuzzleContentPanel();
+		InitializeTouchKeyboardPanel();
+		InitializeTouchArea();
+	}
 
 
 	[Export]
@@ -110,7 +105,7 @@ public class PuzzleSystemGUI : BaseSystemGUI
 	[Export]
 	public NodePath inactivePanelNP;
 
-  [Export]
+	[Export]
 	public NodePath touchAreaNP;
